Still implementing new component system. Fixed lots of issues

+++ include/simulator/base/config.h
@@ -6,10 +6,10 @@
 #define SIMULATOR_CONFIG_H
 
 struct Config {
-    long clock_rate;
-    int process_node;
-    int voltage;
-    int temperature;
+    long clock_rate; // in Hz
+    int process_node; // in nm
+    double voltage;
+    int temperature; // in degrees celsius
 };
 
 #endif  //SIMULATOR_CONFIG_H
